utils: Move process mgmt functions to submodule
