+++ tests/utils/__init__.py
@@ -46,16 +46,6 @@ def test_is_ip(testvalue, expected_result):
     assert expected_result == result
 
 
-def test_is_systemd():
-    # Arrange
-
-    # Act
-    result = utils.is_systemd()
-
-    # Assert
-    assert result
-
-
 def test_get_random_mac(cobbler_api):
     # Arrange
 
@@ -822,59 +812,3 @@ def test_kopts_overwrite():
     # Assert
     assert "textmode" in kopts
     assert "info" in kopts
-
-
-def test_service_restart_no_manager(mocker):
-    # Arrange
-    mocker.patch("cobbler.utils.is_supervisord", autospec=True, return_value=False)
-    mocker.patch("cobbler.utils.is_systemd", autospec=True, return_value=False)
-    mocker.patch("cobbler.utils.is_service", autospec=True, return_value=False)
-
-    # Act
-    result = utils.service_restart("testservice")
-
-    # Assert
-    assert result == 1
-
-
-def test_service_restart_supervisord(mocker):
-    mocker.patch("cobbler.utils.is_supervisord", autospec=True, return_value=True)
-    # TODO Mock supervisor API and return value
-
-    # Act
-    result = utils.service_restart("dhcpd")
-
-    # Assert
-    assert result == 0
-
-
-def test_service_restart_systemctl(mocker):
-    mocker.patch("cobbler.utils.is_supervisord", autospec=True, return_value=False)
-    mocker.patch("cobbler.utils.is_systemd", autospec=True, return_value=True)
-    mocker.patch("cobbler.utils.subprocess_call", autospec=True, return_value=0)
-
-    # Act
-    result = utils.service_restart("testservice")
-
-    # Assert
-    assert result == 0
-    utils.subprocess_call.assert_called_with(
-        ["systemctl", "restart", "testservice"], shell=False
-    )
-
-
-def test_service_restart_service(mocker):
-    # Arrange
-    mocker.patch("cobbler.utils.is_supervisord", autospec=True, return_value=False)
-    mocker.patch("cobbler.utils.is_systemd", autospec=True, return_value=False)
-    mocker.patch("cobbler.utils.is_service", autospec=True, return_value=True)
-    mocker.patch("cobbler.utils.subprocess_call", autospec=True, return_value=0)
-
-    # Act
-    result = utils.service_restart("testservice")
-
-    # Assert
-    assert result == 0
-    utils.subprocess_call.assert_called_with(
-        ["service", "testservice", "restart"], shell=False
-    )
